+++ pass_sc_ebend.h
@@ -3,12 +3,12 @@
 
 #include "statvec.h"
 
-double dx_eBend(short nx, short npx, short nz, short npz, STATvec & sigma, double  r, double iRho, double  Ksc, double  denom);
-double dz_eBend(short nx, short npx, short nz, short npz, STATvec & sigma, double  r, double iRho, double  Ksc, double  denom);
-double dpx_eBend(short nx, short npx, short nz, short npz, STATvec & sigma, double  r, double iRho, double  Ksc, double  denom);
-double dpz_eBend(short nx, short npx, short nz, short npz, STATvec & sigma, double  r, double iRho, double  Ksc, double  denom);
+double dx_eBend(unsigned nx, unsigned npx, unsigned nz, unsigned npz, STATvec & sigma, double  r, double iRho, double  Ksc, double  denom);
+double dz_eBend(unsigned nx, unsigned npx, unsigned nz, unsigned npz, STATvec & sigma, double  r, double iRho, double  Ksc, double  denom);
+double dpx_eBend(unsigned nx, unsigned npx, unsigned nz, unsigned npz, STATvec & sigma, double  r, double iRho, double  Ksc, double  denom);
+double dpz_eBend(unsigned nx, unsigned npx, unsigned nz, unsigned npz, STATvec & sigma, double  r, double iRho, double  Ksc, double  denom);
 
-void sc_eBendPass(STATvec & sigma, double  L, double Angle, double Ksc, short Nint);
+void sc_eBendPass(STATvec & sigma, double  L, double Angle, double Ksc, unsigned Nint);
 
 #endif
